+++ packages/deep-heating/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @home-automation/deep-heating
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - deep-heating-socketio@0.1.1
+
+## 0.0.6
+
+### Patch Changes
+
+- [#1122](https://github.com/GraemeF/home-automation/pull/1122) [`682c648`](https://github.com/GraemeF/home-automation/commit/682c64824eff90b155f06f1c13d7ee0fb396e79e) Thanks [@GraemeF](https://github.com/GraemeF)! - Internal improvements: migrate core systems to Effect, upgrade RxJS to v7, modernise TypeScript configuration, and improve build tooling
+
+- Updated dependencies [[`95b76f4`](https://github.com/GraemeF/home-automation/commit/95b76f4daeaca9fab8f252487769b8a344d6e80a), [`682c648`](https://github.com/GraemeF/home-automation/commit/682c64824eff90b155f06f1c13d7ee0fb396e79e)]:
+  - deep-heating-socketio@0.1.0
+  - @home-automation/deep-heating-web@0.0.2
+
 ## 0.0.5
 
 ### Patch Changes